+++ src/Http/HttpException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Address\Parser\Http;
-
-final class HttpException extends \RuntimeException
-{
-}